requirements.txt
requests==2.7.0
locustio
pyzmq
lxml
pyopenssl
ndg-httpsclient
pyasn1
